fix: free creators blocked from upload + admin dashboard real endpoints
- Fix RequireCreatorTierAttribute: check Role=Creator, not Tier=creator
  Free creators have Tier='free' by default - this was blocking all uploads
- Admin dashboard: replace all stubs with real DB queries
  - GET /admin/tracks - list all tracks with creator info
  - GET /admin/purchases - list all purchases with buyer/track info
  - GET /admin/payouts + /admin/payouts/requests - real payout list
  - POST /admin/payouts/{id}/approve|reject - update payout status + audit log
  - POST /admin/users/{id}/reset-password - generate temp password via Identity
  - GET /admin/settings - returns real TierManifest values (not hardcoded)
  - POST /admin/settings - returns 501 (honestly not implemented)
  - POST /admin/tracks/{id}/feature|pin - return 501 (honestly not implemented)
- Fix TierManifest.Pro.PriceCents: 1499 -> 999 (matches Stripe product at 9.99)
- Block Admin accounts from billing checkout (clear error instead of Stripe 500)
- New DTOs: AdminTrack, AdminPurchase, AdminPayout
- AdminUser DTO: add DisplayName, CreatorTier, UploadCount, CreatedAt
- AdminDashboardSummary: fix count fields from double to int

+++ src/Cambrian.Application/DTOs/Admin/AdminDashboardSummary.cs
@@ -2,13 +2,13 @@ namespace Cambrian.Application.DTOs.Admin;
 
 public class AdminDashboardSummary
 {
-    public double TotalUsers { get; set; }
+    public int TotalUsers { get; set; }
 
-    public double ActiveCreators { get; set; }
+    public int ActiveCreators { get; set; }
 
-    public double TracksUploaded { get; set; }
+    public int TracksUploaded { get; set; }
 
-    public double LicensesSold { get; set; }
+    public int LicensesSold { get; set; }
 
     public double TotalRevenue { get; set; }
